+++ src/components/Project/Hazards/DatasetFlood.tsx
@@ -13,12 +13,12 @@ import validator from "@rjsf/validator-ajv8";
 import config from "@app/app.config";
 
 interface DatasetFloodProps {
-    index: number;
+    value: string;
     projectId: string;
     handleLayerUpdate: (layers: IncoreLayer[]) => void;
 }
 
-export const DatasetFlood: React.FC<DatasetFloodProps> = ({ index, projectId, handleLayerUpdate }) => {
+export const DatasetFlood: React.FC<DatasetFloodProps> = ({ value, projectId, handleLayerUpdate }) => {
     const [loading, setLoading] = useState<boolean>(false);
     const [formKey, setFormKey] = useState<number>(0);
 
@@ -58,7 +58,7 @@ export const DatasetFlood: React.FC<DatasetFloodProps> = ({ index, projectId, ha
     };
 
     return (
-        <TabPanel value={index}>
+        <TabPanel value={value}>
             <Box sx={{ opacity: loading ? 0.5 : 1 }}>
                 <Form
                     key={formKey}