+++ CHANGELOG.md
@@ -8,10 +8,19 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 
 ### Fixed
+
 - Resolve map layer visibility issue when create new hazard [#113](https://github.com/IN-CORE/incore-studio/issues/113)
 
 ### Changed
+
 - Get the layer bounding box and fly to the bound [#124](https://github.com/IN-CORE/incore-studio/issues/124)
+- Allow user to add a region not present in the dropdown [#127](https://github.com/IN-CORE/incore-studio/issues/127)
+
+### Added
+
+- Create Hazard option to project dashboard hazard section [#118](https://github.com/IN-CORE/incore-studio/issues/118)
+- Drawing capability to the map. [#115](https://github.com/IN-CORE/incore-studio/issues/115)
+- Upload dataset modal that uploads dataset to IN-CORE [#128](https://github.com/IN-CORE/incore-studio/issues/128)
 
 ## [Alpha-3] - 03-10-2025